OneDrive - Ensorinstituut/Bureaublad/skynova-website-6TIB/requirements.txt
bcrypt==4.0.1
blinker==1.5
boto3==1.29.2
botocore==1.32.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.3
Flask-Breadcrumbs==0.5.1
Flask-Dropzone==1.6.0
Flask-Mail==0.9.1
Flask-Menu==0.7.2
Flask-PyMongo==2.3.0
Flask-ReCaptcha==0.4.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.0.0
jmespath==1.0.1
MarkupSafe==2.1.2
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
s3transfer==0.7.0
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
stripe==5.2.0
twilio==7.16.4
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
